add logout button
