+++ server/server.js
@@ -96,9 +96,9 @@ app.post(`/notes/:id`, async (request, response) => {
 //DB Connection
 
 if (process.env.NODE_ENV === "production") {
-  app.use(express.static(path.join(__dirname, "client/build")));
+  app.use(express.static(path.join(__dirname, "../client/build")));
   app.get("*", function(req, res) {
-    res.sendFile(path.join(__dirname, "client/build", "index.html"));
+    res.sendFile(path.join(__dirname, "../client/dist", "index.html"));
   });
 }
